all changes

+++ gflib/gpu_regs.h
@@ -8,11 +8,11 @@
 // Exported ROM declarations
 void InitGpuRegManager(void);
 void CopyBufferedValuesToGpuRegs(void);
-void SetGpuReg(u8 regOffset, u16 value);
-void SetGpuReg_ForcedBlank(u8 regOffset, u16 value);
-u16 GetGpuReg(u8 regOffset);
-void SetGpuRegBits(u8 regOffset, u16 mask);
-void ClearGpuRegBits(u8 regOffset, u16 mask);
+void SetGpuReg(u16 regOffset, u16 value);
+void SetGpuReg_ForcedBlank(u16 regOffset, u16 value);
+u16 GetGpuReg(u16 regOffset);
+void SetGpuRegBits(u16 regOffset, u16 mask);
+void ClearGpuRegBits(u16 regOffset, u16 mask);
 void EnableInterrupts(u16 mask);
 void DisableInterrupts(u16 mask);
